Remove vim-geeknote
This plugin no longer appears to be maintained (and I never used it
anyway). Current builds of Vim on Fedora don't include Python 2 support,
so the plugin errors out.

```
line    1:
E319: Sorry, the command is not available in this version: python import sys
line    2:
E319: Sorry, the command is not available in this version: python import vim
line    3:
E319: Sorry, the command is not available in this version: python sys.path.append(vim.eval('expand("<sfile>:h")'))
```