Remove Path header
At one point, I was going to add a link to the plugin from the document,
but then remembered that my use of `mr` means they don't exist in the
repository.

Apparently, while I removed the code that generated the data, I forgot
to remove the heading, which threw off my CSV magic.