Try unsetting mapleader
I've used ',' as my map leader for ages, since it was quick to use.
However, at some point, I started using ';' to jump around lines and I
often wondered if there was an equivalent for moving backwards. Turns
out, I had overridden it. The '\' key is typically fairly large and easy
to hit, and I'm used to using it for both '\' and '|' when coding.